requirements.txt
streamlit==0.86.0
python-dotenv==0.18.0
ibm_watson_machine_learning==1.0.99